# server/src/docs/swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: TomoScan APIs
  description: TomoScan APIs
  license:
    name: Github
    url: 'https://github.com/tomochain/tomoscan'
...
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/accounts:
    get:
      tags:
        - Accounts
      summary: Get list accounts
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/accounts/{hash}:
    get:
      tags:
        - Accounts
      summary: Get account detail
      parameters:
        - name: hash
          in: path
          required: true
          description: account address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/accounts/{hash}/mined:
    get:
      tags:
        - Accounts
      summary: Get list block create
      parameters:
        - name: hash
          in: path
          required: true
          description: account address
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/blocks:
    get:
      tags:
        - Blocks
      summary: Get list blocks
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
  /api/blocks/{hash}:
    get:
      tags:
        - Blocks
      summary: Get block detail
      parameters:
        - name: hash
          in: path
          required: true
          description: block number or block hash
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/blocks/{blockNumber}/address/{address}:
    get:
      tags:
        - Blocks
      summary: Check exist transaction from/to address inside block
      parameters:
        - name: blockNumber
          in: path
          required: true
          description: block number
          type: number
        - name: address
          in: path
          required: true
          description: address
          type: string
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/blocks/signers/{blockNumber}:
    get:
      tags:
        - Blocks
      summary: Get list signers of a block
      parameters:
        - name: blockNumber
          in: path
          required: true
          description: account address
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        500:
          description: Server Internal Error
...
  /api/epochs:
    get:
      tags:
        - Epoch
      summary: Get list epoch
      parameters:
        - name: page
          in: query
          required: false
          description: default = 1
          type: number
        - name: limit
          in: query
          required: false
          description: default 20, maximum = 50
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
  /api/epochs/{epochNumber}:
    get:
      tags:
        - Epoch
      summary: Get epoch detail
      parameters:
        - name: epochNumber
          in: path
          required: true
          description: epoch number
          type: number
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        404:
          description: Not Found
        500:
          description: Server Internal Error
...